update: ratingbar

+++ joekanesiewweek7core2/app/src/main/java/au/swin/joekanesiew_week7_core2/Location.kt
@@ -6,7 +6,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class Location(
     val locationName: String,
-    val locationRating: Int,
+    val locationRating: Float,
     val locationDate: String,
     val locationType: String,
     val locationImage: Int
